requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
google-generativeai
numpy
scikit-learn
requests